made sql dialect not tied to sqlite, ascii builtn added
